Remove dependency on congnito module from config module. Moves cognito pool functions into config.py Enables us to import config into cognito to get config settings.
