+++ cognito.py
@@ -16,91 +16,13 @@ from botocore.exceptions import ClientError, ParamValidationError
 
 from cognito_groups import get_group_by_name, get_group_map
 from logger import LOG
+from config import env_pool_id
 
 CLIENT_EXCEPTIONS = (ClientError, ParamValidationError)
 
 
-def load_app_settings():
-    client = get_boto3_client()
-    pool_id = env_pool_id()
-
-    client_id = ""
-    client_secret = ""
-    cognito_domain = ""
-    estimated_num_users = 0
-
-    pool_client_resp = client.list_user_pool_clients(UserPoolId=pool_id, MaxResults=2)
-
-    if "UserPoolClients" in pool_client_resp:
-        client_id = pool_client_resp["UserPoolClients"][0]["ClientId"]
-
-    if client_id != "":
-        desc_client_resp = client.describe_user_pool_client(
-            UserPoolId=pool_id, ClientId=client_id
-        )
-        if "UserPoolClient" in desc_client_resp:
-            client_secret = desc_client_resp["UserPoolClient"]["ClientSecret"]
-
-    desc_pool = client.describe_user_pool(UserPoolId=pool_id)
-    if "UserPool" in desc_pool:
-        cognito_domain = desc_pool["UserPool"]["Domain"]
-        estimated_num_users = desc_pool["UserPool"]["EstimatedNumberOfUsers"]
-
-    return {
-        "client_id": client_id,
-        "client_secret": client_secret,
-        "cognito_domain": "{}.auth.eu-west-2.amazoncognito.com".format(cognito_domain),
-        "estimated_num_users": estimated_num_users,
-    }
-
-
 def get_boto3_client():
     return boto3.client("cognito-idp", region_name="eu-west-2")
-
-
-def get_cognito_pool_name():
-    environment = os.getenv("APP_ENVIRONMENT", "testing")
-    pool_name_prefix = "corona-cognito-pool-"
-    if environment == "production":
-        pool_name = f"{pool_name_prefix}prod"
-    elif environment == "staging":
-        pool_name = f"{pool_name_prefix}staging"
-    elif environment == "testing":
-        pool_name = f"{pool_name_prefix}development"
-
-    return pool_name
-
-
-def env_pool_id():
-    pool_id = None
-    pool_name = None
-
-    pool_name = get_cognito_pool_name()
-
-    if pool_name is not None:
-        for pool in list_pools():
-            if pool["name"] == pool_name:
-                pool_id = pool["id"]
-                break
-
-    return pool_id
-
-
-def list_pools():
-    client = get_boto3_client()
-    pool_list = []
-    try:
-        response = client.list_user_pools(MaxResults=10)
-    except CLIENT_EXCEPTIONS as error:
-        LOG.error(error)
-        response = {}
-    if "UserPools" in response:
-        # convert keys to lower case
-        pool_list = [
-            {key.lower(): value for key, value in pool.items()}
-            for pool in response["UserPools"]
-        ]
-    return pool_list
 
 
 # TODO remove below once admin app running online
