+++ tests/test_cognito.py
@@ -1,34 +1,7 @@
-import os
-
 import pytest
 
 import cognito
 import stubs
-
-
-def test_env_pool_id_development():
-    user_pool_id = stubs.MOCK_COGNITO_USER_POOL_ID
-    stubber = stubs.mock_cognito_list_pools()
-    with stubber:
-        assert cognito.env_pool_id() == user_pool_id
-
-
-def test_env_pool_id_production(monkeypatch):
-    user_pool_id = stubs.MOCK_COGNITO_USER_POOL_ID
-    monkeypatch.setenv("APP_ENVIRONMENT", "production")
-    stubber = stubs.mock_cognito_list_pools(env="prod")
-    with stubber:
-        assert cognito.env_pool_id() == user_pool_id
-
-
-def test_list_pools():
-    user_pool_id = stubs.MOCK_COGNITO_USER_POOL_ID
-    stubber = stubs.mock_cognito_list_pools()
-
-    with stubber:
-        pools = cognito.list_pools()
-        assert pools[0]["id"] == user_pool_id
-        stubber.deactivate()
 
 
 @pytest.mark.usefixtures("admin_user", "create_user_arguments")
@@ -155,12 +128,3 @@ def test_list_groups_for_user(admin_user):
         first_group = groups["Groups"][0]
         assert first_group["GroupName"] == admin_user["group"]["value"]
         stubber.deactivate()
-
-
-def test_get_cognito_pool_name():
-    os.environ["APP_ENVIRONMENT"] = "production"
-    assert cognito.get_cognito_pool_name() == "corona-cognito-pool-prod"
-    os.environ["APP_ENVIRONMENT"] = "staging"
-    assert cognito.get_cognito_pool_name() == "corona-cognito-pool-staging"
-    os.environ["APP_ENVIRONMENT"] = "testing"
-    assert cognito.get_cognito_pool_name() == "corona-cognito-pool-development"
